feat: watchlist
